+++ scripts/go.mod.cgen
@@ -1,0 +1,59 @@
+module github.com/banzaicloud/istio-operator
+
+go 1.14
+
+require (
+	cloud.google.com/go v0.35.1 // indirect
+	github.com/Masterminds/semver/v3 v3.1.0
+	github.com/banzaicloud/istio-client-go v0.0.9
+	github.com/banzaicloud/k8s-objectmatcher v1.4.0
+	github.com/coreos/go-semver v0.3.0
+	github.com/ghodss/yaml v1.0.0
+	github.com/go-logr/logr v0.1.0
+	github.com/go-logr/zapr v0.1.0 // indirect
+	github.com/gofrs/uuid v3.2.0+incompatible
+	github.com/gogo/protobuf v1.3.1 // indirect
+	github.com/golang/groupcache v0.0.0-20190129154638-5b532d6fd5ef // indirect
+	github.com/google/btree v1.0.0 // indirect
+	github.com/goph/emperror v0.17.2
+	github.com/hashicorp/golang-lru v0.5.0 // indirect
+	github.com/markbates/inflect v1.0.4 // indirect
+	github.com/mattbaird/jsonpatch v0.0.0-20171005235357-81af80346b1a // indirect
+	github.com/mholt/caddy v1.0.0
+	github.com/onsi/gomega v1.5.0
+	github.com/pborman/uuid v1.2.0 // indirect
+	github.com/peterbourgon/diskv v2.0.1+incompatible // indirect
+	github.com/pkg/errors v0.8.1
+	github.com/prometheus/client_golang v0.9.3 // indirect
+	github.com/shurcooL/httpfs v0.0.0-20190707220628-8d4bc4ba7749 // indirect
+	github.com/shurcooL/vfsgen v0.0.0-20200627165143-92b8a710ab6c
+	github.com/spf13/cobra v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/stretchr/testify v1.6.0
+	go.uber.org/atomic v1.4.0 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
+	go.uber.org/zap v1.10.0 // indirect
+	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550 // indirect
+	golang.org/x/net v0.0.0-20191004110552-13f9640d40b9
+	golang.org/x/time v0.0.0-20190308202827-9d24e82272b4
+	golang.org/x/tools v0.0.0-20191119224855-298f0cb1881e // indirect
+	gopkg.in/yaml.v2 v2.2.8 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200603094226-e3079894b1e8 // indirect
+	k8s.io/api v0.17.4
+	k8s.io/apiextensions-apiserver v0.17.4
+	k8s.io/apimachinery v0.17.4
+	k8s.io/client-go v0.17.4
+	k8s.io/kube-openapi v0.0.0-20200410145947-61e04a5be9a6 // indirect
+	sigs.k8s.io/controller-runtime v0.6.0
+	sigs.k8s.io/controller-tools v0.1.9 // indirect
+	sigs.k8s.io/testing_frameworks v0.1.1 // indirect
+)
+
+replace (
+	k8s.io/api => k8s.io/api v0.0.0-20181126151915-b503174bad59
+	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.0.0-20181126155829-0cd23ebeb688
+	k8s.io/apimachinery => k8s.io/apimachinery v0.0.0-20181126123746-eddba98df674
+	k8s.io/client-go => k8s.io/client-go v0.0.0-20181126152608-d082d5923d3c
+	sigs.k8s.io/controller-runtime => sigs.k8s.io/controller-runtime v0.1.9
+	sigs.k8s.io/controller-tools => sigs.k8s.io/controller-tools v0.1.9
+)
